# docs/swagger.yaml
basePath: /
definitions:
  models.RequestBody:
    properties:
      number:
        example: 65
        type: integer
    type: object
  models.Response:
    properties:
      message:
        example: Saved!
        type: string
    type: object
  models.ResponseCollection:
    properties:
      message:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
    type: object
  models.ResponseGet:
    properties:
      message:
        example: Type 1
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A Multipliers Service API in Go using Gin framework
  title: Multipliers
  version: "1.0"
paths:
  /healthcheck:
    get:
      description: Checks if the service is up and returns the string "UP"
      produces:
      - application/json
      responses:
        "200":
          description: UP
          schema:
            type: string
      summary: Checks if the service is up
      tags:
      - Healtcheck
  /multipliers:
    get:
      description: Returns the value from a specific number between 1 and 100 if the
        number is multiplier from 3, 5 or both returns Type 1, Type 2 or Type 3 respectively
      parameters:
      - description: Un numero entre 1 y 100
        in: query
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseGet'
      summary: Retrieves a value from a specific number
      tags:
      - Get a value
    post:
      description: Receives a number and saves it into a collection
      parameters:
      - description: A number between 1-100
        in: body
        name: number
        required: true
        schema:
          $ref: '#/definitions/models.RequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Saves a number into a collection
      tags:
      - Save
  /multipliers/collection:
    get:
      description: Returns the values from a collection, if its empty returns null
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseCollection'
      summary: Retrieves the values from a collection
      tags:
      - Get collection
swagger: "2.0"